Prototype Binding composition
This simplifies binding the TextBoxes in the UI.  Using composition a single function may be called for each.  That function composes a binding for the controls Text and IsEnabled properties as well as attaching to WPF's TextChanged event for the purpose of forwarding the notifications to INotifyingObject.

Much of this will migrate to RabidWarren.Binding.